convert desk structure to support sanity v3
